Test for duplicates only in test code, not in production code
Disable the duplicate check on WDI which is being worked on.
